requirements.txt
 streamlit 
 csv 
time
 matplotlib.pyplot 
seaborn 


